requirements.txt
edge_tts==7.2.3
fastapi==0.116.1
mistralai==1.9.9
protobuf==6.32.0
pydantic==2.11.7
python-dotenv==1.1.1
whisper==1.1.10
